move fix_zip from zip handler to unicorn zip handler
